remplacement du h1 de carte par un message indiquant que le panier est vide

+++ front/js/confirmation.js
@@ -2,7 +2,21 @@
 let url = new URL(window.location.href);
 let orderId = url.searchParams.get("orderId");
 
-const orderNumber = document.querySelector('#orderId');
-orderNumber.textContent = orderId;
+if (orderId === null || isCartEmpty() ) {
+    let emptyCart = document.querySelector('.confirmation');
+    emptyCart.textContent = "Erreur";
+} else {
 
-localStorage.clear();
+    const orderNumber = document.querySelector('#orderId');
+    orderNumber.textContent = orderId;
+    localStorage.clear();
+    //
+}
+
+
+
+function isCartEmpty() {
+    const cart = JSON.parse(localStorage.getItem('cart'));
+    return !cart || cart.length === 0;
+}
+
